feat: change creaate and update

+++ RLPortalBackend/Services/Impl/CourseSectionService.cs
@@ -1,8 +1,7 @@
 ﻿using AutoMapper;
-using Microsoft.OpenApi.Writers;
 using RLPortalBackend.Entities;
+using RLPortalBackend.Exceptions;
 using RLPortalBackend.Models.CourseSection;
-using RLPortalBackend.Models.Test;
 using RLPortalBackend.Models.Theory;
 using RLPortalBackend.Repositories;
 
@@ -13,15 +12,15 @@ namespace RLPortalBackend.Services.Impl
 
         private readonly ICourseSectionRepository _courseSectionRepository;
         private readonly IMapper _mapper;
-        private readonly ITestService _testService;
         private readonly ITheoryRepository _theoryRepository;
+        private readonly ITestRepository _testRepository;
 
-        public CourseSectionService(ICourseSectionRepository courseSectionRepository, IMapper mapper, ITestService testService, ITheoryRepository theoryRepository)
+        public CourseSectionService(ICourseSectionRepository courseSectionRepository, IMapper mapper, ITheoryRepository theoryRepository, ITestRepository testRepository)
         {
             _courseSectionRepository = courseSectionRepository;
             _mapper = mapper;
-            _testService = testService;
             _theoryRepository = theoryRepository;
+            _testRepository = testRepository;
         }
 
 
@@ -29,52 +28,67 @@ namespace RLPortalBackend.Services.Impl
         public async Task<CourseSectionDto> CreateAsync(NewCourseSectionDto newCourseSectionDto)
         {
             CourseSectionEntity courseSectionEntity = _mapper.Map<CourseSectionEntity>(newCourseSectionDto);
-
-            ICollection<NoIdTheoryDto> theories = newCourseSectionDto.TheoryDtos;
-            ICollection<TheoryEntity> theoryEntities = _mapper.Map<ICollection<TheoryEntity>>(theories);
-
-            await _theoryRepository.CreateManyAsync(theoryEntities);
-
-            ICollection<Guid> theoryGuids = theoryEntities.Select(e => e.Id).ToList();
-            courseSectionEntity.TheoryEntityId = theoryGuids;
-
-            ICollection<Guid> testGuids = new List<Guid>();
-            ICollection<CreateTest> tests = newCourseSectionDto.TestDtos;
-            foreach (CreateTest test in tests)
-            {
-                var dtoWithId = await _testService.CreateAsync(test);
-                testGuids.Add(dtoWithId.Id);
-            }
-            courseSectionEntity.TestEntityId = testGuids;
-
+            var testId = Guid.NewGuid();
+            var theoryId = Guid.NewGuid();
+            TestEntity testEntity = new(testId, "", new List<Guid>());
+            await _testRepository.CreateAsync(testEntity);
+            await _theoryRepository.CreateAsync(new TheoryEntity(theoryId, "", "", "", new List<TheorySectionEntity>()));
+            courseSectionEntity.TestEntityId = testId;
+            courseSectionEntity.TheoryEntityId = theoryId;
             await _courseSectionRepository.CreateAsync(courseSectionEntity);
-
             CourseSectionDto dto = _mapper.Map<CourseSectionDto>(courseSectionEntity);
+            dto.TheoryDto = _mapper.Map<TheoryDto>(await _theoryRepository.GetAsync(theoryId));
             return dto;
+
         }
 
         public async Task<ICollection<CourseSectionDto>> GetAsync()
         {
             ICollection<CourseSectionEntity> courseSectionEntities = await _courseSectionRepository.GetAsync();
-            ICollection<CourseSectionDto> dtos = _mapper.Map<ICollection<CourseSectionDto>>(courseSectionEntities);
+            ICollection<CourseSectionDto> dtos = new List<CourseSectionDto>();
+            foreach(var courseSectionEntity in courseSectionEntities)
+            {
+                var dto = _mapper.Map<CourseSectionDto>(courseSectionEntity);
+                dto.TheoryDto = _mapper.Map<TheoryDto>(await _theoryRepository.GetAsync(courseSectionEntity.TheoryEntityId));
+                dtos.Add(dto);
+            }
+            //ICollection<CourseSectionDto> dtos = _mapper.Map<ICollection<CourseSectionDto>>(courseSectionEntities);
             return dtos;
         }
 
         public async Task<CourseSectionDto> GetByIdAsync(Guid id)
         {
+            if (await _courseSectionRepository.GetAsync(id) == null)
+            {
+                throw new CourseSectionNotFoundException($"Course section {id} not found");
+            }
             CourseSectionEntity courseSectionEntity = await _courseSectionRepository.GetAsync(id);
             CourseSectionDto dto = _mapper.Map<CourseSectionDto>(courseSectionEntity);
+            dto.TheoryDto = _mapper.Map<TheoryDto>(await _theoryRepository.GetAsync(courseSectionEntity.TheoryEntityId));
             return dto;
         }
 
         public async Task RemoveAsync(Guid id)
         {
+            if (await _courseSectionRepository.GetAsync(id) == null)
+            {
+                throw new CourseSectionNotFoundException($"Course section {id} not found");
+            }
             await _courseSectionRepository.RemoveAsync(id);
         }
 
         public async Task UpdateAsync(Guid id, NewCourseSectionDto newCourseSectionDto)
         {
+            if (await _courseSectionRepository.GetAsync(id) == null)
+            {
+                throw new CourseSectionNotFoundException($"Course section {id} not found");
+            }
             CourseSectionEntity courseSectionEntity = _mapper.Map<CourseSectionEntity>(newCourseSectionDto);
+            var entity = await _courseSectionRepository.GetAsync(id);
+            courseSectionEntity.TheoryEntityId = entity.TheoryEntityId;
+            courseSectionEntity.TestEntityId = entity.TestEntityId;
+            courseSectionEntity.Id = id;
+            await _courseSectionRepository.UpdateAsync(id, courseSectionEntity);
 
         }
